requirements.txt
Django==3.1.5
django-simpleui==2021.1.1
mysqlclient==2.0.1
requests==2.25.1
djangorestframework==3.12.2
coreapi==2.3.3
websocket-client==0.58.0

